requirements.txt
accelerate==0.33.0
aiobotocore==2.22.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.12.0
aiosignal==1.3.2
anndata==0.11.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
botocore==1.37.3
cellxgene-census==1.15.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses==0.6
datasets==2.14.4
debugpy==1.8.11
decorator==5.1.1
deepchem==2.8.0
defusedxml==0.7.1
descriptastorus==2.8.0
dill==0.3.7
e3fp==1.2.7
et_xmlfile==2.0.0
evaluate==0.4.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.58.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
fuzzywuzzy==0.18.0
gdown==5.2.0
gget==0.29.1
h11==0.14.0
h5py==3.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.10
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
legacy-api-wrap==1.4.1
littleutils==0.2.4
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.0.2
mmh3==5.1.0
mordredcommunity==2.0.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
mysql-connector-python==9.3.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ogb==1.3.6
omegaconf==2.3.0
openpyxl==3.1.5
outdated==0.2.2
overrides==7.7.0
packaging==24.2
pandas==1.5.3
pandas_flavor==0.7.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.8
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
PySocks==1.7.1
pytdc==1.1.15
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rdkit==2023.9.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
s3fs==2024.12.0
safetensors==0.4.5
scanpy==1.11.2
scikit-fingerprints==1.16.0
scikit-learn==1.7.0
scipy==1.13.1
sdaxen-python-utilities==0.1.5
seaborn==0.13.2
Send2Trash==1.8.3
session-info2==0.1.2
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
somacore==1.0.11
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tiledb==0.29.1
tiledbsoma==1.11.4
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.1.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.4.0
xxhash==3.5.0
yarl==1.18.3
torch_geometric==2.6.1

